core-plugin-api: apply extension fixes and deprecate extensions in core-app-api
