+++ .changeset/olive-singers-sparkle.md
@@ -1,0 +1,5 @@
+---
+'@backstage/core-plugin-api': patch
+---
+
+Apply fixes to the extension creation API that were mistakenly applied to `@backstage/core-app-api` instead.
